칭호테이블 목표 챌린지개수 필드추가

+++ src/main/java/com/greeneyback/member/dto/RstrntDTO.java
@@ -20,7 +20,7 @@ public class RstrntDTO {
     private String rstrntMenuinfo;
     private String rstrntLa;
     private String rstrntLo;
-    private Integer rstrntStar;
+    private float rstrntStar;
     private Integer areaCode;
 
     public static RstrntDTO toRstrntDTO(RstrntEntity rstrntEntity) {
